extract canvas logic in top level layout
