+++ packages/toolpad-app/src/runtime/index.ts
@@ -1,0 +1,2 @@
+export { default } from './ToolpadApp';
+export * from './ToolpadApp';
